+++ classroom.rb
@@ -1,12 +1,14 @@
 class Classroom
-    attr_accessor :label
-    def initialize(label)
-        @label = label
-        @students = []
-    end
+  attr_accessor :label
+  attr_reader :students
 
-    def add_student(student)
-        @students.push(student)
-        student.classroom = self
+  def initialize(label)
+    @label = label
+    @students = []
+  end
 
+  def add_student(student)
+    @students.push(student)
+    student.classroom = self
+  end
 end
