+++ person.rb
@@ -2,7 +2,7 @@ require './nameable'
 
 class Person < Nameable
   attr_accessor :name, :age
-  attr_reader :id
+  attr_reader :id, :rentals
 
   def initialize(age, name = 'Unknown', parent_permission: true)
     super()
@@ -24,6 +24,7 @@ class Person < Nameable
   def add_rental(rental)
     @rentals.push(rental)
     rental.person = self
+  end
 
   private
