--- retnal.rb
@@ -1,11 +1,0 @@
-class Rental
-    attr_accessor :date
-    attr_reader :book, :person
-    def initialize(date)
-        @date = date
-        @book = book
-        book.rentals << self
-        @person = person
-        person.rentals << self
-    end
-end
