+++ student.rb
@@ -1,6 +1,9 @@
 require './person'
+require './classroom'
+
 class Student < Person
   attr_reader :classroom
+
   def initialize(classroom, age, name = 'Unknown', parent_permission: true)
     super(age, name, parent_permission)
     @classroom = classroom
@@ -13,5 +16,5 @@ class Student < Person
   def classroom=(classroom)
     @classroom = classroom
     classroom.students.push(self) unless classroom.students.include?(self)
-end
+  end
 end
